+ Initial port for windows platform + mvc compiller support + replace hunter with conan + minor refractoring ang bug fixes

+++ src/common/Exception.cpp
@@ -7,7 +7,7 @@ namespace Common {
 
     Exception::Exception(const char *fmt, ...) {
         va_list args;
-        int sizeBuffer = strlen(fmt) * 2;
+        size_t sizeBuffer = strlen(fmt) * 2;
         std::string str;
 
         while (true) {
